fix(auth-portal): Track Auth API Events More Consistently
We now consistently add tracking events for the auth-api events we need to trigger pipeline from
