+++ bin/auth-api/src/routes/workspace.routes.ts
@@ -1,11 +1,9 @@
-import _ from "lodash";
 import { nanoid } from "nanoid";
 import { z } from "zod";
 import { InstanceEnvType } from "@prisma/client";
 import { ApiError } from "../lib/api-error";
 import { getCache, setCache } from "../lib/cache";
 import {
-  getUserByEmail,
   getUserById,
   refreshUserAuth0Profile,
 } from "../services/users.service";
@@ -28,7 +26,7 @@ import { validate } from "../lib/validation-helpers";
 import { CustomRouteContext } from "../custom-state";
 import { createSdfAuthToken } from "../services/auth.service";
 import { tracker } from "../lib/tracker";
-import { extractAdminAuthUser, extractAuthUser, router } from ".";
+import { extractAuthUser, router } from ".";
 
 router.get("/workspaces", async (ctx) => {
   const authUser = extractAuthUser(ctx);
@@ -36,7 +34,7 @@ router.get("/workspaces", async (ctx) => {
 });
 
 // :workspaceId named param handler - little easier for TS this way than using router.param
-async function extractWorkspaceIdParam(ctx: CustomRouteContext) {
+export async function extractWorkspaceIdParam(ctx: CustomRouteContext) {
   if (!ctx.params.workspaceId) {
     throw new Error(
       "Only use this fn with routes containing :workspaceId param",
@@ -70,106 +68,18 @@ router.get("/workspaces/:workspaceId", async (ctx) => {
 });
 
 router.delete("/workspaces/:workspaceId", async (ctx) => {
+  const authUser = extractAuthUser(ctx);
   const workspace = await extractOwnWorkspaceIdParam(ctx);
 
   await deleteWorkspace(workspace.id);
 
+  tracker.trackEvent(authUser, "workspace_deleted", {
+    workspaceId: workspace.id,
+    workspaceDeletedAt: new Date(),
+    workspaceDeletedBy: authUser.email,
+  });
+
   ctx.body = "";
-});
-
-export type WorkspaceLookup = {
-  firstName?: string | null;
-  lastName?: string | null;
-  email?: string | null;
-  displayName: string;
-  instanceUrl: string | null;
-};
-router.get("/workspaces/admin-lookup/:workspaceId", async (ctx) => {
-  // Just for authorization, result is discarded
-  extractAdminAuthUser(ctx);
-
-  const workspace = await extractWorkspaceIdParam(ctx);
-  const user = await getUserById(workspace.creatorUserId);
-
-  const workspaceDetails: WorkspaceLookup = {
-    firstName: user?.firstName,
-    lastName: user?.lastName,
-    email: user?.email,
-    displayName: workspace.displayName,
-    instanceUrl: workspace.instanceUrl,
-  };
-
-  ctx.body = workspaceDetails;
-});
-
-router.post("/workspaces/setup-production-workspace", async (ctx) => {
-  // Just for authorization, result is discarded
-  extractAdminAuthUser(ctx);
-
-  const reqBody = validate(
-    ctx.request.body,
-    z.object({
-      userEmail: z.string(),
-    }),
-  );
-
-  const user = await getUserByEmail(reqBody.userEmail);
-  if (user) {
-    const userWorkspaces = await getUserWorkspaces(user.id);
-    const hasDefaultWorkspace = _.head(
-      _.filter(
-        userWorkspaces,
-        (w) => w.isDefault && w.creatorUserId === user.id,
-      ),
-    );
-
-    const workspaceDetails = await createWorkspace(
-      user,
-      InstanceEnvType.SI,
-      SAAS_WORKSPACE_URL,
-      `${user.nickname}'s Production Workspace`,
-      hasDefaultWorkspace === null || hasDefaultWorkspace === undefined,
-    );
-
-    ctx.body = {
-      newWorkspace: workspaceDetails,
-    };
-  }
-});
-
-router.post("/workspaces/setup-production-workspace-by-userid", async (ctx) => {
-  // Just for authorization, result is discarded
-  extractAdminAuthUser(ctx);
-
-  const reqBody = validate(
-    ctx.request.body,
-    z.object({
-      userId: z.string(),
-    }),
-  );
-
-  const user = await getUserById(reqBody.userId);
-  if (user) {
-    const userWorkspaces = await getUserWorkspaces(user.id);
-    const hasDefaultWorkspace = _.head(
-      _.filter(
-        userWorkspaces,
-        (w) => w.isDefault && w.creatorUserId === user.id,
-      ),
-    );
-
-    const workspaceDetails = await createWorkspace(
-      user,
-      InstanceEnvType.SI,
-      SAAS_WORKSPACE_URL,
-      `${user.nickname}'s Production Workspace`,
-      hasDefaultWorkspace === null || hasDefaultWorkspace === undefined,
-    );
-
-    ctx.body = {
-      newWorkspace: workspaceDetails,
-    };
-  }
 });
 
 router.post("/workspaces/new", async (ctx) => {
@@ -227,44 +137,11 @@ router.patch("/workspaces/:workspaceId", async (ctx) => {
     workspace.quarantinedAt,
   );
 
-  ctx.body = await getUserWorkspaces(authUser.id);
-});
-
-router.patch("/workspaces/:workspaceId/quarantine", async (ctx) => {
-  const authUser = extractAdminAuthUser(ctx);
-
-  const workspace = await extractWorkspaceIdParam(ctx);
-
-  const reqBody = validate(
-    ctx.request.body,
-    z.object({
-      isQuarantined: z.boolean(),
-    }),
-  );
-
-  const quarantineDate = new Date();
-  if (reqBody.isQuarantined) {
-    tracker.trackEvent(authUser, "quarantine_workspace", {
-      quarantinedBy: authUser.email,
-      quarantinedAt: quarantineDate,
-      workspaceId: workspace.id,
-    });
-  } else {
-    tracker.trackEvent(authUser, "unquarantine_workspace", {
-      unQuarantinedBy: authUser.email,
-      unQuarantinedAt: quarantineDate,
-      workspaceId: workspace.id,
-    });
-  }
-
-  const quarantinedAt = reqBody.isQuarantined ? quarantineDate : null;
-
-  await patchWorkspace(
-    workspace.id,
-    workspace.instanceUrl,
-    workspace.displayName,
-    quarantinedAt,
-  );
+  tracker.trackEvent(authUser, "workspace_updated", {
+    workspaceId: workspace.id,
+    workspaceUpdatedAt: new Date(),
+    workspaceUpdatedBy: authUser.email,
+  });
 
   ctx.body = await getUserWorkspaces(authUser.id);
 });
@@ -311,7 +188,7 @@ router.post("/workspace/:workspaceId/membership", async (ctx) => {
     }),
   );
 
-  tracker.trackEvent(authUser, "change_workspace_member_role", {
+  tracker.trackEvent(authUser, "workspace_membership_roles_changed", {
     role: reqBody.role,
     userId: reqBody.userId,
     workspaceId: workspace.id,
@@ -337,7 +214,7 @@ router.post("/workspace/:workspaceId/membership", async (ctx) => {
 
 router.post("/workspace/:workspaceId/members", async (ctx) => {
   // user must be logged in
-  extractAuthUser(ctx);
+  const authUser = extractAuthUser(ctx);
 
   const workspace = await extractOwnWorkspaceIdParam(ctx);
 
@@ -348,7 +225,7 @@ router.post("/workspace/:workspaceId/members", async (ctx) => {
     }),
   );
 
-  await inviteMember(reqBody.email, workspace.id);
+  await inviteMember(authUser, reqBody.email, workspace.id);
 
   const members: Member[] = [];
   const workspaceMembers = await getWorkspaceMembers(workspace.id);
@@ -363,12 +240,18 @@ router.post("/workspace/:workspaceId/members", async (ctx) => {
     });
   });
 
+  tracker.trackEvent(authUser, "workspace_user_invited", {
+    workspaceId: workspace.id,
+    memberAdded: reqBody.email,
+    memberAddedAt: new Date(),
+  });
+
   ctx.body = members;
 });
 
 router.delete("/workspace/:workspaceId/members", async (ctx) => {
   // user must be logged in
-  extractAuthUser(ctx);
+  const authUser = extractAuthUser(ctx);
 
   const workspace = await extractOwnWorkspaceIdParam(ctx);
 
@@ -391,6 +274,12 @@ router.delete("/workspace/:workspaceId/members", async (ctx) => {
       role: wm.roleType,
       signupAt: wm.user.signupAt,
     });
+  });
+
+  tracker.trackEvent(authUser, "workspace_user_removed", {
+    workspaceId: workspace.id,
+    memberRemoved: reqBody.email,
+    memberRemovedAt: new Date(),
   });
 
   ctx.body = members;
